feat: set input page as home
