--- src/lib/components/Nav.svelte
@@ -1,6 +1,0 @@
-<nav>
-  <a href="/">Dashboard</a> |
-  <a href="/input">Input</a> |
-  <a href="/compare">Compare</a> |
-  <a href="/export">Export</a>
-</nav>
