+++ src/lib/components/index.ts
@@ -1,5 +1,4 @@
 import DataTable from './DataTable.svelte';
-import Nav from './Nav.svelte';
 
-export { DataTable, Nav };
+export { DataTable };
 export default DataTable;
